feat: move slimewars in to it's own seperate project to leave Game as a clean set of base functionality separate from any implementation.
