+++ Games.sln
@@ -15,6 +15,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AutomatedPlayer", "src\Auto
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{67C74B82-9C52-43D4-8EB5-636B514798EE}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Games", "Games", "{B638AA7A-35F1-4341-911C-5056A11F454D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SlimeWars", "src\Games\SlimeWars\SlimeWars.csproj", "{BFF781A8-C1D9-4E22-A08A-0BC9EF8BB447}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +45,10 @@ Global
 		{2BB92683-0A3B-47FD-BB90-CEF9068CB86A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2BB92683-0A3B-47FD-BB90-CEF9068CB86A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2BB92683-0A3B-47FD-BB90-CEF9068CB86A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BFF781A8-C1D9-4E22-A08A-0BC9EF8BB447}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BFF781A8-C1D9-4E22-A08A-0BC9EF8BB447}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BFF781A8-C1D9-4E22-A08A-0BC9EF8BB447}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BFF781A8-C1D9-4E22-A08A-0BC9EF8BB447}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -51,6 +59,8 @@ Global
 		{54ED0759-94AA-4E8D-8CCC-F33836D5080C} = {67C74B82-9C52-43D4-8EB5-636B514798EE}
 		{411AA2D4-C50D-472F-B11E-CC52C1C916F1} = {67C74B82-9C52-43D4-8EB5-636B514798EE}
 		{2BB92683-0A3B-47FD-BB90-CEF9068CB86A} = {67C74B82-9C52-43D4-8EB5-636B514798EE}
+		{B638AA7A-35F1-4341-911C-5056A11F454D} = {67C74B82-9C52-43D4-8EB5-636B514798EE}
+		{BFF781A8-C1D9-4E22-A08A-0BC9EF8BB447} = {B638AA7A-35F1-4341-911C-5056A11F454D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {FA0DF96A-1F22-403E-BBE4-5F6FB8B8F69D}
