+++ src/Game/Player.cs
@@ -1,10 +1,6 @@
 ﻿using Common.Interface;
 using System;
-using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Game
 {
