+++ src/WebApp/Pages/Index.razor
@@ -1,6 +1,6 @@
 ﻿@using ComponentLibrary.Board
 @using ComponentLibrary.WinLoseDraw
-@using Game
+@using SlimeWars
 @using AutomatedPlayer
 @using Common.Enum;
 @page "/"
